複数サービスを切り替えるための基盤構築
