+++ CalcLib/CalcLib.csproj
@@ -41,9 +41,16 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CalcButton.cs" />
+    <Compile Include="Maeda\Basis\SvcHolder.cs" />
+    <Compile Include="Maeda\CalcuratorSvc.cs" />
+    <Compile Include="Maeda\Basis\IBackendSvc.cs" />
+    <Compile Include="Maeda\Basis\SvcTree.cs" />
+    <Compile Include="Maeda\Dispatcher\SvcDispatcher.cs" />
+    <Compile Include="Maeda\OmikujiSvc.cs" />
     <Compile Include="Maeda\CalcSvcMaeda.cs" />
     <Compile Include="CalcSvc.cs" />
     <Compile Include="CalcContext.cs" />
+    <Compile Include="Maeda\Util\SvcActivator.cs" />
     <Compile Include="Yamamoto\CalcSvcYamamoto.cs" />
     <Compile Include="Moriguchi\CalcSvcMoriguchi.cs" />
     <Compile Include="Factory.cs" />
@@ -52,6 +59,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Yamamoto\DecimalExt.cs" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
